+++ nodalia-cards.manifest.js
@@ -1,8 +1,8 @@
 export default {
-  "pkgVersion": "1.1.0-alpha.4",
-  "contentSha256_12": "b86a65a8e582",
+  "pkgVersion": "1.1.0-alpha.6",
+  "contentSha256_12": "be7973b707f9",
   "file": "nodalia-cards.bundle.js"
 };
-export const pkgVersion = "1.1.0-alpha.4";
-export const contentSha256_12 = "b86a65a8e582";
+export const pkgVersion = "1.1.0-alpha.6";
+export const contentSha256_12 = "be7973b707f9";
 export const file = "nodalia-cards.bundle.js";